+++ src/Point.tsx
@@ -1,4 +1,5 @@
 import * as React from 'react';
+import initialProps from './initialProps';
 import { IInternalPointProps } from './utils/interfaces';
 
 export default function Point({ unit, point, index, moveStartHandler, moveEndHandler}: {unit: string, point: IInternalPointProps, index: number, moveStartHandler: CallableFunction, moveEndHandler: CallableFunction}) {
@@ -10,6 +11,7 @@ export default function Point({ unit, point, index, moveStartHandler, moveEndHan
       onMouseUp={(e) => moveEndHandler(e, index)}
       onMouseDown={(e) => moveStartHandler(e, index)}
       style={{
+        backgroundColor: initialProps.defaultBackgroundColorOfPoint,
         height: point.height,
         transform: `translate3d(${(point.translateX || 0) + 0}${unit}, ${(point.translateY || 0) + 0}${unit}, 0)`,
         width: point.width,