+++ src/initialProps.ts
@@ -1,5 +1,6 @@
 export default {
   axis: 'xy',
+  defaultBackgroundColorOfPoint: 'red',
   height: 'auto',
   multiple: false,
   onReach: (e: any) => null,